refactor: 546 - change course entity re-exports
