+++ src/entities/course/index.ts
@@ -1,2 +1,3 @@
-export type { Course, CourseItem, CourseStatus } from './types';
+export type { Course, CourseItemData, CourseStatus } from './types';
 export { CourseCard } from './ui/course-card/course-card';
+export { CourseItem } from './ui/course-item/course-item';
