+++ src/entities/course/ui/course-item/course-item.test.tsx
@@ -1,9 +1,8 @@
 import { screen } from '@testing-library/react';
-import { CourseItem as TCourseItem } from '@/entities/course';
-import { CourseItem } from '@/entities/course/ui/course-item/course-item.tsx';
+import { CourseItem, CourseItemData } from '@/entities/course';
 import { renderWithRouter } from '@/shared/__tests__/utils';
 
-const mockedProps: TCourseItem = {
+const mockedProps: CourseItemData = {
   title: 'Introduction to React',
   language: ['en'],
   startDate: '2024-05-01',
